pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-pyoc"
version = "0.12.0"
description = "封装一些 Python 小工具，提升开发效率。"
authors = [
    { name = "one-ccs", email = "one-ccs@foxmail.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "typing-extensions>=4.14.1",
]

[project.optional-dependencies]
qt5 = [
    "PyQt5",
]
flask = [
    "flask",
]
yaml = [
    "PyYAML",
]

[project.urls]
homepage = "https://github.com/one-ccs/easy_pyoc"
repository = "https://github.com/one-ccs/easy_pyoc.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["easy_pyoc"]

[dependency-groups]
dev = [
    "flask>=3.1.2",
    "pytest>=9.0.2",
]
